requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
langchain==0.1.0
langchain-community==0.0.10
langchain-groq>=0.0.1
langgraph==0.0.20
groq==0.4.1
faiss-cpu==1.7.4
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.0
numpy==1.24.3
pandas==2.1.3
matplotlib==3.8.2
seaborn==0.13.0
python-dotenv==1.0.0
pydantic==2.5.0
redis==5.0.1
prometheus-client==0.19.0
scikit-learn==1.3.2
rouge-score==0.1.2
nltk==3.8.1

